Further cleanup of jq related issues.
deploy_utils.sh
===============
Add fail_if_null() function that can be used to fail if a specified
variable that was initialised with the result of a jq query has the
value 'null', which happens when you query an non-existent entry.

Add error_msg() function that can be used to report error messages
in a consistent fashion.

TODO:
  * Moved colourised message support into deploy_utils.sh and have
    error_msg() output errors in red.

validate.sh
===========
The validate.sh script's usage of jq doesn't correctly handle the
case where we get a 'null' back when attempting to lookup an invalid
reference.

Additionally in a number of cases the checks for values being set in
the config files appear to have been mistakenly inverted, checking for
retrieve values being an empty, rather than non-empty, string. This
seems to have "worked" previously due to the the 'null' value being
returned, and thus the inverted check appeared to do the right thing.

Additionally I added support for tracking if errors occurred during
the validating checking, allowing the validate.sh command to exit
with a non-zero exit status indicating failure.  The error() function
makes this easy to manage as well as printing the error message in
bold red text.

The new heading() function prints headings in cyan with a following
line of '-' characters in a consistent fashion, with consistent
formatting/layout.

Fixed a whitespace issue in the indentation of the displayed subnet
address range for the deployer.

TODO:
  * Convert the repeated pattern of code for the subnet processing
    when validating sap_system config files into a routine.
  * Convert the output echo commands to use printf commands with
    format strings the handle the whitespace indentation.

prepare_region.sh
=================
Updates jq queries to use --raw-output and removes post processing
that is no longer required.

Also catch any non-zero exit status, rather than specifically 255,
when calling other scripts and exit out immediately with that exit
status.

install_deployer.sh, remove_region.sh, remover.sh
====================================================================
Updates jq queries to use --raw-output and removes post processing
that is no longer required.

install_library.sh
============
Leverages fail_if_null() to fail when jq queries result in a 'null'
value.

Also updates jq queries to use --raw-output and removes the post
processing that is no longer required.

installer.sh
============
Ensire that if we get a 'null' back when checking for .deployer.use
that we treat it as a false.

Leverages fail_if_null() to fail when jq queries result in a 'null'
value.

Also updates jq queries to use --raw-output and removes the post
processing that is no longer required.

Also quotes all jq query strings that contain shell pattern match
characters to avoid potential for side effects.

install_workloadzone.sh
=======================
Simplifies and logically fixes some if conditions to simply check
for values being empty or non-empty value, rather than a negated
non-empty or empty value, respectively. In some cases we were
checking if a value was non-empty, and if so dynamically retrieved
the correct value again.

Leverages fail_if_null() to fail when jq queries result in a 'null'
value.

Also updates jq queries to use --raw-output and removes the post
processing that is no longer required.

Also quotes all jq query strings that contain shell pattern match
characters to avoid potential for side effects.