Ansible inventory and ansible.cfg settings
Update the generated inventory to include an ansible_user setting so
that we don't have to specify the remote user to use when running any
ansible commands.

Configure the default inventory, private_key_file settings so that is
not necessary to specify them when running any ansible commands.

Note that while the private_key_file name will be specified in the
ansible.cfg, the specified file won't exist until it is downloaded
from the appropriate storage account keyvault.