Use cloud-init config to disable growpart
By defining a cloud-init config that disables running growpart to
expand the root file system we have the option to manage how the
extra space on the OS disk is utilised later.

Currently this is required for SUSE distro images which leverage
cloud-init to handle this automation root filesystem expansion.

A new cloudinit_disable_growpart variable is defined in the
common_infrastructure whose value can be used as the custom_data
attribute when defining a Linux VM. The other relevant modules
have been updated to leverage this variable when being initialised.

All Linux VM definitions, except for the Deployer, have had the
custom_data attribute added to them that leverages this new
cloudinit_disable_growpart value.