Ensure correct IP address used for HDB inventory entries
In recent testing I attempted to bring up a HDB node backed by a RHEL
for SAP Apps image, but was unable to talk to the node using Ansible
via the IP address specified in the terraform generated inventory.

After discussing with the team it was identified that the inventory
was being generated with the "admin" subnet IP address for the HDB
nodes, rather than the correct DB subnet address; this patch corrects
that error.