Confirm environment and region non-empty in set_secrets.sh
In order to safely construct the path to the relevant config file
we need to ensure that the environment and region values that are
provider via CLI option, or via prompted user input, are non-empty
values that conform to the expected validation pattern.

Also when prompting for the user to enter a password field we don't
need to explicitly manage terminal echoing if we are specifying the
-s option to the read command since it will both handle disablement
of terminal echo correctly, and more importantly, will restore the
terminal echo mode if the user Ctrl-C's the script run.

Additionally the -e option enables readline processing which makes
no sense when enterring a password, as we don't necessarily want
or need readline editing features when entering a password value.