Enhance region and environment regexp checks
Define valid_environment and valid_region_name helper functions in
deploy_utils.sh that can be used to validate that environment and
region name strings conform to the expected formats.

As a starting point we are requiring the following:

  * environments are specified as a string of at most 5 characters,
    and at least 2 characters, starting with an uppercase letter,
    followed by one or more letters or numbers.

  * region names are specified as a string of lowercase letters
    optionally followed by a number.

Update set_secrets.sh to leverage these new helper functions when
checking that provided environment and region values are specified
correctly.